requirements.txt
pytest
pytest-bdd
playwright
pytest-playwright
allure-pytest-bdd
